change sort key to double precision

+++ src/main/java/org/zaikorea/ZaiClient/request/CartaddEvent.java
@@ -12,7 +12,7 @@ public class CartaddEvent extends Event {
     public CartaddEvent(String userId, String itemId, double timestamp) {
         this.userId = userId;
         this.itemId = itemId;
-        this.timestamp = (float)timestamp;
+        this.timestamp = timestamp;
         this.eventType = defaultEventType;
         this.eventValue = defaultEventValue;
     }
