+++ src/main/java/org/zaikorea/ZaiClient/request/LikeEvent.java
@@ -12,7 +12,7 @@ public class LikeEvent extends Event {
     public LikeEvent(String userId, String itemId, double timestamp) {
         this.userId = userId;
         this.itemId = itemId;
-        this.timestamp = (float)timestamp;
+        this.timestamp = timestamp;
         this.eventType = defaultEventType;
         this.eventValue = defaultEventValue;
     }
